move some helpers from command package to cli
