+++ src/cmd/cli/command/deploymentinfo.go
@@ -19,32 +19,3 @@ func printPlaygroundPortalServiceURLs(serviceInfos []*defangv1.ServiceInfo) {
 		}
 	}
 }
-
-func printEndpoints(serviceInfos []*defangv1.ServiceInfo) {
-	for _, serviceInfo := range serviceInfos {
-		andEndpoints := ""
-		if len(serviceInfo.Endpoints) > 0 {
-			andEndpoints = "and will be available at:"
-		}
-
-		serviceConditionText := "has status " + serviceInfo.Status
-		if serviceInfo.State != defangv1.ServiceState_NOT_SPECIFIED {
-			serviceConditionText = "is in state " + serviceInfo.State.String()
-		}
-
-		term.Info("Service", serviceInfo.Service.Name, serviceConditionText, andEndpoints)
-		for i, endpoint := range serviceInfo.Endpoints {
-			if serviceInfo.Service.Ports[i].Mode == defangv1.Mode_INGRESS {
-				endpoint = "https://" + endpoint
-			}
-			term.Println("   -", endpoint)
-		}
-		if serviceInfo.Domainname != "" {
-			if serviceInfo.ZoneId != "" {
-				term.Println("   -", "https://"+serviceInfo.Domainname)
-			} else {
-				term.Println("   -", "https://"+serviceInfo.Domainname+" (after `defang cert generate` to get a TLS certificate)")
-			}
-		}
-	}
-}
