+++ src/cmd/cli/command/deploymentinfo_test.go
@@ -33,36 +33,3 @@ func TestPrintPlaygroundPortalServiceURLs(t *testing.T) {
 		t.Errorf("got %q, want %q", got, want)
 	}
 }
-
-func TestPrintEndpoints(t *testing.T) {
-	defaultTerm := term.DefaultTerm
-	t.Cleanup(func() {
-		term.DefaultTerm = defaultTerm
-	})
-
-	var stdout, stderr bytes.Buffer
-	term.DefaultTerm = term.NewTerm(os.Stdin, &stdout, &stderr)
-
-	printEndpoints([]*defangv1.ServiceInfo{
-		{
-			Service: &defangv1.Service{
-				Name: "service1",
-				Ports: []*defangv1.Port{
-					{Mode: defangv1.Mode_INGRESS},
-					{Mode: defangv1.Mode_HOST},
-				},
-			},
-			Status: "UNKNOWN",
-			Endpoints: []string{
-				"example.com",
-				"service1.internal",
-			},
-		}})
-	const want = ` * Service service1 has status UNKNOWN and will be available at:
-   - https://example.com
-   - service1.internal
-`
-	if got := stdout.String(); got != want {
-		t.Errorf("got %q, want %q", got, want)
-	}
-}
